+++ src/main/java/kitchenpos/domain/MenuProduct.java
@@ -1,41 +1,66 @@
 package kitchenpos.domain;
 
+import java.util.Objects;
 
+import javax.persistence.CascadeType;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+
+@Entity
+@Table(name = "menu_product")
 public class MenuProduct {
-    private Long seq;
-    private Long menuId;
-    private Long productId;
-    private long quantity;
 
-    public Long getSeq() {
-        return seq;
-    }
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long seq;
 
-    public void setSeq(final Long seq) {
-        this.seq = seq;
-    }
+	@ManyToOne(cascade = CascadeType.ALL)
+	@JoinColumn(name = "menu_id")
+	private Menu menu;
 
-    public Long getMenuId() {
-        return menuId;
-    }
+	@ManyToOne(cascade = CascadeType.ALL)
+	@JoinColumn(name = "product_id")
+	private Product product;
 
-    public void setMenuId(final Long menuId) {
-        this.menuId = menuId;
-    }
+	private long quantity;
 
-    public Long getProductId() {
-        return productId;
-    }
+	protected MenuProduct() {
+	}
 
-    public void setProductId(final Long productId) {
-        this.productId = productId;
-    }
+	public MenuProduct(Menu menu, Product product, long quantity) {
+		setMenu(menu);
+		setProduct(product);
+		this.quantity = quantity;
+	}
 
-    public long getQuantity() {
-        return quantity;
-    }
+	public Long getProductId() {
+		return product.getId();
+	}
 
-    public void setQuantity(final long quantity) {
-        this.quantity = quantity;
-    }
+	public long getQuantity() {
+		return quantity;
+	}
+
+	public void setMenu(Menu newMenu) {
+		if (Objects.nonNull(menu)) {
+			menu.getMenuProducts().remove(this);
+		}
+		menu = newMenu;
+		if (!menu.getMenuProducts().contains(this)) {
+			menu.getMenuProducts().add(this);
+		}
+	}
+
+	public void setProduct(Product newProduct) {
+		this.product = newProduct;
+	}
+
+	public Money getPurchasePrice() {
+		return product.getPrice().multiply(quantity);
+	}
 }
